Changed TODO related things in org mode
Added adtional suequence of keywords for things that need to be worked
out within a project. Simply to mark a header NEEDSWORK so it's
differnt then TODO.

Checkboxes where replaced with icons so we don't want emacs to draw
boxes

For the TODO sequence we want to see the words. An icon for NEEDSWORK