Fix to exclude flyspell mode from certain buffers
mode-name can sometimes be a list (e.g. emacs lisp buffer)
This fixed the issue by checking if mode-name is a list and then only
comparing the first element of the list. It seems that the first
element is a string. at least in the case of emacs lisp buffer.